+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* split deployment manifests (#11) 
 * Project Status: beta phase (#10)
 * Remove multizone and fix URL environment (#4)